Reintroduce custom eigendecomposition
Note however, due to synchronization issues with the GPU kernel function
since the algorithm is not entirely parallelized, we're limited to a
maximum matrix size of 32x32 (aka 1024 threads per block).

We might be able to overcome this limitation by using intra-block
synchronization...

+++ include/geigen/geigen.h
@@ -14,7 +14,7 @@ struct eigensystem
   eigensystem(std::vector<T> values, std::vector<T> vectors) : values(values), vectors(vectors) {}
 };
 
-eigensystem<float> compute_eigensystem(const std::vector<float>& matrix, const int n);
-eigensystem<float> compute_eigensystem_magma(const std::vector<float>& matrix, const int n);
+eigensystem<float> compute_eigensystem(const std::vector<float>& matrix, const size_t n);
+eigensystem<float> compute_eigensystem_magma(const std::vector<float>& matrix, const size_t n);
 
 }
